requirements.txt
django == 3.0.3
django-markdownx
gunicorn == 20.0.4
libsass
django-compressor
django-sass-processor
pillow